requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.7.3; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
chardet==3.0.4
discord.py==1.6.0
idna==3.1; python_version >= '3.4'
multidict==5.1.0; python_version >= '3.6'
python-environ==0.4.54
typing-extensions==3.7.4.3
yarl==1.6.3; python_version >= '3.6'
